+++ core/CHANGELOG.md
@@ -5,3 +5,9 @@
   - 🆕 `components/actorGenerator` v1.0.0.
   - 🆕 `components/functionalities/hitTheWallFunctionalityBuilder` v1.0.0.
   - 🆕 `components/functionalities/positioningFunctionalityBuilder` v1.0.0.
+  - 🆕 `engineGenerator` v1.0.0.
+  - 🆕 `engineGenerator/processors/processorGenerator` v1.0.0.
+  - 🆕 `engineGenerator/processors/positioningProcessorBuilder` v1.0.0.
+  - 🆕 `engineGenerator/processors/checkers/engineErrorBuilder` v1.0.0.
+  - 🆕 `engineGenerator/processors/checkers/hasHitAWallBuilder` v1.0.0.
+  - 🆕 `engineGenerator/processors/checkers/hasHitAWallErrorGenerator` v1.0.0.